+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changelog
 
+## 2.3.0 (2021-02-21)
+
+- Reflect API changes:
+  - Now `Transaction` has `attachments: List<Attachment>` and `labels: List<Label>` fields - `attachmentIds`
+    and `labelIds` are therefore now deprecated
+  - A new `getTransaction(id)` API to get a single `Transaction`
+- Add a `bypassSslChecks` option on `HttpConfiguration` (should never be used in production)
+- Update all dependencies to their latest version
+
 ## 2.2.0 (2020-12-12)
+
 - Updated all dependencies to their latest version
 - Use the new qonto.com domain instead of qonto.eu
 - Reflect API changes (a few new fields and objects)
 
 ## 2.1.0
+
 Add the `fileName: String` field to the `Attachment` model.
 
 ## v2.0.1